+++ azurerm/internal/services/storage/parse/account_id_test.go
@@ -1,4 +1,4 @@
-package parsers
+package parse
 
 import (
 	"testing"
