+++ azurerm/internal/services/storage/parse/storage_container_data_plane.go
@@ -1,4 +1,4 @@
-package parsers
+package parse
 
 import (
 	"fmt"
