+++ azurerm/internal/services/storage/parse/storage_container_resource_manager_test.go
@@ -1,4 +1,4 @@
-package parsers
+package parse
 
 import (
 	"testing"
