+++ azurerm/internal/services/storage/parse/storage_sync.go
@@ -1,4 +1,4 @@
-package parsers
+package parse
 
 import (
 	"fmt"
