+++ azurerm/internal/services/storage/resource_arm_storage_data_lake_gen2_filesystem.go
@@ -10,7 +10,7 @@ import (
 	"github.com/hashicorp/terraform-plugin-sdk/helper/schema"
 	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/helpers/tf"
 	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/internal/clients"
-	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/internal/services/storage/parsers"
+	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/internal/services/storage/parse"
 	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/internal/timeouts"
 	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/utils"
 	"github.com/tombuildsstuff/giovanni/storage/2019-12-12/datalakestore/filesystems"
@@ -77,7 +77,7 @@ func resourceArmStorageDataLakeGen2FileSystemCreate(d *schema.ResourceData, meta
 	ctx, cancel := timeouts.ForCreate(meta.(*clients.Client).StopContext, d)
 	defer cancel()
 
-	storageID, err := parsers.ParseAccountID(d.Get("storage_account_id").(string))
+	storageID, err := parse.ParseAccountID(d.Get("storage_account_id").(string))
 	if err != nil {
 		return err
 	}
@@ -132,7 +132,7 @@ func resourceArmStorageDataLakeGen2FileSystemUpdate(d *schema.ResourceData, meta
 		return err
 	}
 
-	storageID, err := parsers.ParseAccountID(d.Get("storage_account_id").(string))
+	storageID, err := parse.ParseAccountID(d.Get("storage_account_id").(string))
 	if err != nil {
 		return err
 	}
@@ -172,7 +172,7 @@ func resourceArmStorageDataLakeGen2FileSystemRead(d *schema.ResourceData, meta i
 		return err
 	}
 
-	storageID, err := parsers.ParseAccountID(d.Get("storage_account_id").(string))
+	storageID, err := parse.ParseAccountID(d.Get("storage_account_id").(string))
 	if err != nil {
 		return err
 	}
