+++ azurerm/internal/services/storage/resource_arm_storage_sync.go
@@ -12,7 +12,7 @@ import (
 	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/helpers/tf"
 	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/internal/clients"
 	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/internal/location"
-	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/internal/services/storage/parsers"
+	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/internal/services/storage/parse"
 	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/internal/services/storage/validate"
 	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/internal/tags"
 	azSchema "github.com/terraform-providers/terraform-provider-azurerm/azurerm/internal/tf/schema"
@@ -28,7 +28,7 @@ func resourceArmStorageSync() *schema.Resource {
 		Delete: resourceArmStorageSyncDelete,
 
 		Importer: azSchema.ValidateResourceIDPriorToImport(func(id string) error {
-			_, err := parsers.ParseStorageSyncID(id)
+			_, err := parse.ParseStorageSyncID(id)
 			return err
 		}),
 
@@ -117,7 +117,7 @@ func resourceArmStorageSyncRead(d *schema.ResourceData, meta interface{}) error 
 	ctx, cancel := timeouts.ForRead(meta.(*clients.Client).StopContext, d)
 	defer cancel()
 
-	id, err := parsers.ParseStorageSyncID(d.Id())
+	id, err := parse.ParseStorageSyncID(d.Id())
 	if err != nil {
 		return err
 	}
@@ -145,7 +145,7 @@ func resourceArmStorageSyncUpdate(d *schema.ResourceData, meta interface{}) erro
 	ctx, cancel := timeouts.ForUpdate(meta.(*clients.Client).StopContext, d)
 	defer cancel()
 
-	id, err := parsers.ParseStorageSyncID(d.Id())
+	id, err := parse.ParseStorageSyncID(d.Id())
 	if err != nil {
 		return err
 	}
@@ -179,7 +179,7 @@ func resourceArmStorageSyncDelete(d *schema.ResourceData, meta interface{}) erro
 	ctx, cancel := timeouts.ForDelete(meta.(*clients.Client).StopContext, d)
 	defer cancel()
 
-	id, err := parsers.ParseStorageSyncID(d.Id())
+	id, err := parse.ParseStorageSyncID(d.Id())
 	if err != nil {
 		return err
 	}
