+++ azurerm/internal/services/storage/tests/resource_arm_storage_data_lake_gen2_filesystem_test.go
@@ -8,7 +8,7 @@ import (
 	"github.com/hashicorp/terraform-plugin-sdk/terraform"
 	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/internal/acceptance"
 	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/internal/clients"
-	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/internal/services/storage/parsers"
+	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/internal/services/storage/parse"
 	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/utils"
 )
 
@@ -115,7 +115,7 @@ func testCheckAzureRMStorageDataLakeGen2FileSystemExists(resourceName string) re
 		}
 
 		fileSystemName := rs.Primary.Attributes["name"]
-		storageID, err := parsers.ParseAccountID(rs.Primary.Attributes["storage_account_id"])
+		storageID, err := parse.ParseAccountID(rs.Primary.Attributes["storage_account_id"])
 		if err != nil {
 			return err
 		}
@@ -143,7 +143,7 @@ func testAzureRMStorageDataLakeGen2StorageAccountDelete(resourceName string) res
 			return fmt.Errorf("Not found: %s", resourceName)
 		}
 
-		storageID, err := parsers.ParseAccountID(rs.Primary.Attributes["storage_account_id"])
+		storageID, err := parse.ParseAccountID(rs.Primary.Attributes["storage_account_id"])
 		if err != nil {
 			return err
 		}
@@ -166,7 +166,7 @@ func testCheckAzureRMStorageDataLakeGen2FileSystemDestroy(s *terraform.State) er
 		}
 
 		fileSystemName := rs.Primary.Attributes["name"]
-		storageID, err := parsers.ParseAccountID(rs.Primary.Attributes["storage_account_id"])
+		storageID, err := parse.ParseAccountID(rs.Primary.Attributes["storage_account_id"])
 		if err != nil {
 			return err
 		}
