+++ azurerm/internal/services/storage/validate/storage_container.go
@@ -3,7 +3,7 @@ package validate
 import (
 	"fmt"
 
-	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/internal/services/storage/parsers"
+	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/internal/services/storage/parse"
 )
 
 func StorageContainerResourceManagerID(i interface{}, k string) (warnings []string, errors []error) {
@@ -13,7 +13,7 @@ func StorageContainerResourceManagerID(i interface{}, k string) (warnings []stri
 		return
 	}
 
-	if _, err := parsers.StorageContainerResourceManagerID(v); err != nil {
+	if _, err := parse.StorageContainerResourceManagerID(v); err != nil {
 		errors = append(errors, fmt.Errorf("Can not parse %q as a resource id: %v", k, err))
 		return
 	}
